// src/api/models/business-creation-model.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Argento Oficina Virtual
 * Api para Portal de Clientes Argento
 *
 * The version of the OpenAPI document: 0.10.2
 * Contact: info@mts.com.do
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { BusinessIdentificationType } from './business-identification-type';
// May contain unused imports in some cases
// @ts-ignore
import { BusinessType } from './business-type';
// May contain unused imports in some cases
// @ts-ignore
import { ContactInfoModel } from './contact-info-model';

/**
 * 
 * @export
 * @interface BusinessCreationModel
 */
export interface BusinessCreationModel {
    /**
     * 
     * @type {string}
     * @memberof BusinessCreationModel
     */
    'name'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof BusinessCreationModel
     */
    'code'?: string | null;
    /**
     * 
     * @type {BusinessIdentificationType}
     * @memberof BusinessCreationModel
     */
    'identificationType'?: BusinessIdentificationType;
    /**
     * 
     * @type {string}
     * @memberof BusinessCreationModel
     */
    'identification'?: string | null;
    /**
     * 
     * @type {BusinessType}
     * @memberof BusinessCreationModel
     */
    'type'?: BusinessType;
    /**
     * 
     * @type {any}
     * @memberof BusinessCreationModel
     */
    'dataByType'?: any | null;
    /**
     * 
     * @type {string}
     * @memberof BusinessCreationModel
     */
    'responsibleUserId'?: string | null;
    /**
     * 
     * @type {Array<ContactInfoModel>}
     * @memberof BusinessCreationModel
     */
    'contactInfo'?: Array<ContactInfoModel> | null;
    /**
     * 
     * @type {string}
     * @memberof BusinessCreationModel
     */
    'addressLine1'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof BusinessCreationModel
     */
    'addressLine2'?: string | null;
    /**
     * 
     * @type {number}
     * @memberof BusinessCreationModel
     */
    'ownerId'?: number | null;
    /**
     * 
     * @type {boolean}
     * @memberof BusinessCreationModel
     */
    'inheritServiceRequestTypesFromOwner'?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof BusinessCreationModel
     */
    'assignAllServiceRequestTypes'?: boolean;
}