// src/api/models/business-edition-model.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Argento Oficina Virtual
 * Api para Portal de Clientes Argento
 *
 * The version of the OpenAPI document: 0.10.2
 * Contact: info@mts.com.do
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { BusinessIdentificationType } from './business-identification-type';
// May contain unused imports in some cases
// @ts-ignore
import { BusinessType } from './business-type';
// May contain unused imports in some cases
// @ts-ignore
import { ContactInfoModel } from './contact-info-model';
// May contain unused imports in some cases
// @ts-ignore
import { Status } from './status';

/**
 * 
 * @export
 * @interface BusinessEditionModel
 */
export interface BusinessEditionModel {
    /**
     * 
     * @type {string}
     * @memberof BusinessEditionModel
     */
    'name'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof BusinessEditionModel
     */
    'code'?: string | null;
    /**
     * 
     * @type {BusinessIdentificationType}
     * @memberof BusinessEditionModel
     */
    'identificationType'?: BusinessIdentificationType;
    /**
     * 
     * @type {string}
     * @memberof BusinessEditionModel
     */
    'identification'?: string | null;
    /**
     * 
     * @type {BusinessType}
     * @memberof BusinessEditionModel
     */
    'type'?: BusinessType;
    /**
     * 
     * @type {any}
     * @memberof BusinessEditionModel
     */
    'dataByType'?: any | null;
    /**
     * 
     * @type {string}
     * @memberof BusinessEditionModel
     */
    'responsibleUserId'?: string | null;
    /**
     * 
     * @type {Array<ContactInfoModel>}
     * @memberof BusinessEditionModel
     */
    'contactInfo'?: Array<ContactInfoModel> | null;
    /**
     * 
     * @type {string}
     * @memberof BusinessEditionModel
     */
    'addressLine1'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof BusinessEditionModel
     */
    'addressLine2'?: string | null;
    /**
     * 
     * @type {number}
     * @memberof BusinessEditionModel
     */
    'id'?: number;
    /**
     * 
     * @type {Status}
     * @memberof BusinessEditionModel
     */
    'status'?: Status;
}