// src/api/models/user-type.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Argento Oficina Virtual
 * Api para Portal de Clientes Argento
 *
 * The version of the OpenAPI document: 0.10.2
 * Contact: info@mts.com.do
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @enum {string}
 */

export const UserType = {
    Business: 'Business',
    Admin: 'Admin',
    SystemAdmin: 'SystemAdmin'
} as const;

export type UserType = typeof UserType[keyof typeof UserType];